refactor 피드백반영 - 로직에 따른 개행 반영 - 매직 리터럴 추가 - LottoResult 패키지 구조 변경 - LottoResultTest (LottoResult 패키지 구조 변경에 따른 수정)
