+++ src/test/java/lotto/LottoResultTest.java
@@ -1,6 +1,5 @@
 package lotto;
 
-import lotto.domain.LottoResult;
 import lotto.domain.LottoTicket;
 import lotto.domain.Lottos;
 import lotto.domain.WinningNumbers;
